Add some realistic notification tests.
Currently about half of these are failing.

They don't all pass, most are minor things like writing more inboxes than required. One failure would be a privacy bug except that another ACL check later on catches it. Two of them are currently unresearched but seem to be related to sending one less email than expected in some cases.

Another two are actually behavior I don't want (but others may argue about): when you comment on an entry in a channel it shouldn't be shown to all your followers, just like when you post to a channel it is only shown to the people in the channel. Ideally those comments would only be shown to people in the channel who are also your follower. I expect Ade has opinions on this.

**Update** Submitting for now testing current behavior, we can hash out changes later

Review URL: http://rietku.appspot.com/40001

git-svn-id: file:///home/robert/svn/p/trunk@137 e2b98fe4-98ed-42b4-beac-2e5b131153b4